requirements.txt
streamlit>=1.34.0
plotly==5.19.0
pandas==2.2.1
numpy==1.26.4
dhanhq==2.0.1
python-dotenv==1.0.1
